backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15.0,<3.16
djangorestframework-simplejwt>=5.3.1,<5.4
django-cors-headers>=4.3.1,<4.4
django-environ>=0.11.2,<0.12
psycopg[binary]>=3.1,<3.2
drf-spectacular>=0.27.2,<0.28
